requirements.txt
streamlit==1.20.0  # Ensure compatibility with Streamlit Cloud
pandas==1.5.2
numpy==1.24.4
matplotlib==3.6.3
scikit-learn==1.3.1
Pillow==9.5.0
diffusers==0.23.0  # Compatible with huggingface-hub
accelerate==0.23.0
ftfy==6.1.1
seaborn==0.13.0
huggingface-hub==0.17.0  # Downgrade to ensure compatibility with tokenizers
torch==1.12.0  
kaggle
tensorflow

